Refactor Parser (#28)
* Parser interface, dependencies udpated

* Get options moved to internal

* Internal/util.go refactor and tests
