+++ internal/options.go
@@ -3,6 +3,7 @@ package internal
 import (
 	"encoding/json"
 	"errors"
+	"flag"
 	"net/http"
 	"strings"
 )
@@ -16,6 +17,20 @@ type Options struct {
 	Init       bool
 	Quiet      bool
 	Version    bool
+}
+
+func GetOptions() Options {
+	var opts Options
+
+	flag.BoolVar(&opts.ClearCache, "no-cache", false, "Clear Goke's cache. Default: false")
+	flag.BoolVar(&opts.Watch, "watch", false, "Goke remains on and watches the task's specified files for changes, then reruns the command. Default: false")
+	flag.BoolVar(&opts.Force, "force", false, "Executes the task regardless whether the files have changed or not. Default: false")
+	flag.BoolVar(&opts.Init, "init", false, "Initializes a goke.yml file in the current directory")
+	flag.BoolVar(&opts.Quiet, "quiet", false, "Disables all output to the console. Default: false")
+	flag.BoolVar(&opts.Version, "version", false, "Prints the current Goke version")
+	flag.Parse()
+
+	return opts
 }
 
 func (opts *Options) InitHandler() error {
